Separate serialized class to file
